requirements.txt
colorama==0.3.9
decorator==4.1.2
diff-match-patch==20121119
Django==1.11.6
djangorestframework==3.7.1
et-xmlfile==1.0.1
ipython==6.2.1
ipython-genutils==0.2.0
jdcal==1.3
jedi==0.11.0
odfpy==1.3.5
openpyxl==2.4.9
parso==0.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
